cleanup the structure
