+++ csv/post-repliesDelaware.csv
@@ -1,0 +1,1 @@
+-0.2,Dynamism,1,Neutral,1
