+++ csv/post-repliesOhio.csv
@@ -1,0 +1,13 @@
+-100.0,Neutral,123,Neutral,148
+-100.0,Gratefulness,3,Amusement,11
+-69.5,Dishonesty,15,Relief,1
+-68.69,Dynamism,3,Sadness,5
+-61.59,Bias,4,Disgust,5
+-61.59,Unlawfulness,10,Joy,1
+-60.5,Self-restraint,1,Love,1
+-54.9,Disagreement,2,Excitement,2
+-52.79,Awareness,3,Shame,2
+-50.5,Initiative,1,Surprise,2
+-49.9,Unawareness,1,Compassion,1
+-48.7,Violence,2,Guilt,1
+-48.4,Honesty,3,Happiness,1
