+++ csv/post-repliesWest Virginia.csv
@@ -1,0 +1,10 @@
+-100.0,Neutral,135,Neutral,172
+-100.0,Gratefulness,1,Surprise,5
+-100.0,Addiction,5,Amusement,4
+-95.19,Dynamism,3,Disgust,1
+-88.0,Bias,4,Anger,2
+-88.0,Unlawfulness,6,Irritation,1
+-79.19,Violence,29,Sadness,1
+-74.8,Humour,3,Hatred,1
+-64.0,Ignorance,1,Excitement,1
+-61.59,Irresponsibility,2,Embarrassment,1
