--- post-repliesCalifornia.csv
@@ -1,13 +1,0 @@
--100.0,Neutral,125,Neutral,148
--100.0,Gratefulness,3,Amusement,11
--69.5,Dishonesty,15,Relief,1
--68.69,Bias,2,Sadness,5
--61.59,Unlawfulness,8,Disgust,5
--61.59,Self-restraint,1,Joy,1
--60.5,Disagreement,2,Love,1
--54.9,Awareness,3,Excitement,2
--52.79,Initiative,1,Shame,2
--50.5,Unawareness,1,Surprise,2
--49.9,Violence,2,Compassion,1
--48.7,Honesty,3,Guilt,1
--48.4,Commitment,1,Happiness,1
