--- post-repliesNew Jersey.csv
@@ -1,1 +1,0 @@
--15.39,Neutral,2,Neutral,3
