--- post-repliesNew York.csv
@@ -1,177 +1,0 @@
-sent,emot,emot_f,behav,behav_f
--100.0,Neutral,139,Compassion,1
--100.0,Concern,3,Neutral,153
--100.0,Dishonesty,8,Affection,2
--100.0,Competence,4,Repulsion,1
--100.0,Curiosity,1,Amusement,11
--98.59,Bias,2,Fear,1
--83.59,Lack of intelligence,3,Sadness,1
--52.79,Rejection,1,Regret,1
--52.79,Neglect,1,Surprise,4
--41.09,Healthy lifestyle,2,Embarrassment,1
--40.0,Apprehension,1,Worry,1
--35.5,Gluttony,1,Relief,1
--35.2,Inexperience,1,,
--30.79,Violence,2,,
--29.39,Unlawfulness,2,,
--29.39,Passivity,2,,
--29.2,Impoliteness,1,,
--27.7,Humour,3,,
--26.39,Dissoluteness,1,,
--23.39,Gratefulness,1,,
--23.29,Initiative,2,,
--18.7,Positivity,1,,
--18.6,Dynamism,1,,
--17.6,,,,
--15.39,,,,
--13.5,,,,
--12.8,,,,
--12.39,,,,
--12.39,,,,
--12.19,,,,
--11.19,,,,
--10.6,,,,
--10.3,,,,
--10.0,,,,
--9.3,,,,
--9.0,,,,
--8.8,,,,
--8.8,,,,
--8.3,,,,
--8.19,,,,
--7.5,,,,
--6.59,,,,
--5.8,,,,
--4.5,,,,
--4.5,,,,
--4.09,,,,
--3.9,,,,
--2.79,,,,
--2.59,,,,
--2.2,,,,
--1.39,,,,
--1.1,,,,
--0.8,,,,
--0.4,,,,
--0.4,,,,
--0.2,,,,
--0.2,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.0,,,,
-0.89,,,,
-1.1,,,,
-1.29,,,,
-1.7,,,,
-1.79,,,,
-1.79,,,,
-1.89,,,,
-2.0,,,,
-2.09,,,,
-2.09,,,,
-2.29,,,,
-2.29,,,,
-2.5,,,,
-2.59,,,,
-2.7,,,,
-2.7,,,,
-3.2,,,,
-3.2,,,,
-3.29,,,,
-4.19,,,,
-4.3,,,,
-4.3,,,,
-4.8,,,,
-6.0,,,,
-6.09,,,,
-6.3,,,,
-6.4,,,,
-6.4,,,,
-7.8,,,,
-9.6,,,,
-9.6,,,,
-9.69,,,,
-10.19,,,,
-10.89,,,,
-10.89,,,,
-10.89,,,,
-11.6,,,,
-11.89,,,,
-12.8,,,,
-12.8,,,,
-12.8,,,,
-13.1,,,,
-13.8,,,,
-16.2,,,,
-16.39,,,,
-16.39,,,,
-16.39,,,,
-16.39,,,,
-18.29,,,,
-19.2,,,,
-19.2,,,,
-19.2,,,,
-19.5,,,,
-21.89,,,,
-23.7,,,,
-24.79,,,,
-25.6,,,,
-26.0,,,,
-26.29,,,,
-27.39,,,,
-27.39,,,,
-27.39,,,,
-32.9,,,,
-32.9,,,,
-46.59,,,,
-54.79,,,,
-54.79,,,,
-93.3,,,,
