+++ templates/bar_chart.html
