--- web_modified/requirements.txt
@@ -1,14 +1,0 @@
-Flask
-WTForms
-SQLAlchemy
-bcrypt
-requests
-flask-heroku
-gunicorn
-matplotlib
-jinja2
-pandas
-numpy
-matplotlib
-Jinja2
-expertai-nlapi
